chore(security): safely fetch the sourced files contents before sourcing them
This prevents the execution of malicious altered code on sourced in
default scripts.

+++ extra/docker/clear-cache-soft.sh
@@ -4,7 +4,10 @@
 # Clears docker build and volume cache
 
 # shellcheck source=extra/_error.sh
-\. "./../_error.sh"
+if ! e=$(util-fetch "$(realpath "./../_error.sh" || true)"); then
+  exit 1
+fi
+\. <(echo "$e")
 
 docker builder prune
 docker volume prune
